+++ src/RGBA32Image.cpp
@@ -1,7 +1,7 @@
 #include "RGBA32Image.hpp"
 #include <cassert>
 
-cpp_renderer::RGBA32Image::RGBA32Image(unsigned int width, unsigned int height,
+cpp_renderer::RGBA32Image::RGBA32Image(size_t width, size_t height,
                                        RGBA32Color background_color) {
   width_ = width;
   height_ = height;
@@ -15,8 +15,8 @@ cpp_renderer::RGBA32Image::RGBA32Image(unsigned int width, unsigned int height,
   }
 }
 
-cpp_renderer::RGBA32Color
-cpp_renderer::RGBA32Image::getPixel(unsigned int x, unsigned int y) const {
+cpp_renderer::RGBA32Color cpp_renderer::RGBA32Image::getPixel(size_t x,
+                                                              size_t y) const {
   assert(x < width_);
   assert(y < height_);
   RGBA32Color color;
@@ -25,7 +25,7 @@ cpp_renderer::RGBA32Image::getPixel(unsigned int x, unsigned int y) const {
   return color;
 }
 
-void cpp_renderer::RGBA32Image::setPixel(unsigned int x, unsigned int y,
+void cpp_renderer::RGBA32Image::setPixel(size_t x, size_t y,
                                          RGBA32Color color) {
   assert(x < width_);
   assert(y < height_);
@@ -37,11 +37,11 @@ const uint8_t *cpp_renderer::RGBA32Image::getRawData() const {
   return data_.data();
 }
 
-unsigned int cpp_renderer::RGBA32Image::getWidth() const { return width_; }
+size_t cpp_renderer::RGBA32Image::getWidth() const { return width_; }
 
-unsigned int cpp_renderer::RGBA32Image::getHeight() const { return height_; }
+size_t cpp_renderer::RGBA32Image::getHeight() const { return height_; }
 
-void cpp_renderer::RGBA32Image::resize(unsigned int width, unsigned int height,
+void cpp_renderer::RGBA32Image::resize(size_t width, size_t height,
                                        RGBA32Color background_color) {
   width_ = width;
   height_ = height;
